// src/entities/core-entities/enums/status-description.ts
export enum StatusDescription {
  None = 0,
  HT = 1,
  OTHT = 2,
  HomeRetired = 3,
  AwayRetired = 4,
  LostCoverage = 5,
  MedicalTimeout = 6,
  TimeoutHomeTeam = 7,
  TimeoutAwayTeam = 8,
  Timeout = 9,
  HomeWalkover = 10,
  AwayWalkover = 11,
  MatchDrawn = 12,
  BallMarkInspection = 13,
  BallMarkInspectionCompleted = 14,
  CorrectionMode = 15,
  MedicalTreatment = 16,
  TreatmentCompleted = 17,
  PlayersOnCourt = 18,
  TossStarted = 19,
  TossSaved = 20,
  UmpireOnCourt = 21,
  UmpiringCancelled = 22,
  WarmUpStarted = 23,
  WarmUpCompleted = 24,
  MedicalTimeoutCancelled = 25,
  MedicalTimeoutCompleted = 26,
  SwapServer = 27,
  SwapReceiver = 28,
  ChallengeStarted = 29,
  ChallengeCompleted = 30,
  ChallengeCancelled = 31,
  ToiletBreakStarted = 32,
  ToiletBreakCompleted = 33,
  ToiletBreakCancelled = 34,
  EvaluationStarted = 35,
  TimePenalty = 36,
  GamePenalty = 37,
  CoachingStarted = 38,
  CoachingCancelled = 39,
  CoachingCompleted = 40,
  Breaktime = 41,
  PointStarted = 42,
  PointScored = 43,
  PointFault = 44,
  PointReplayed = 45,
}